Upstream update

+++ chirp/drivers/kguv8dplus.py
@@ -666,9 +666,9 @@ class KGUV8DPlusRadio(chirp_common.CloneModeRadio,
         _nam = self._memobj.names[number]
 
         if mem.empty:
-            _mem.set_raw("\x00" * (_mem.size() / 8))
+            _mem.set_raw("\x00" * int((_mem.size() / 8)))
             self._memobj.valid[number] = 0
-            self._memobj.names[number].set_raw("\x00" * (_nam.size() / 8))
+            self._memobj.names[number].set_raw("\x00" * int((_nam.size() / 8)))
             return
 
         _mem.rxfreq = int(mem.freq / 10)
